Model_Item
修改错误的存储类型(Json不能解析float;类型的数据)

+++ Assets/Scripts/Model/Prop/Model_Item.cs
@@ -39,7 +39,7 @@ public class Model_Item
 
     public Drawing[] drawing;//制作所需物品
     public int drawingItemId;//设计图对应的ID
-    public float makeTime;//制作需要的时间
+    public int makeTime;//制作需要的时间
 
     public class Drawing
     {
@@ -56,7 +56,7 @@ public class Model_Item
         int sellPriceByGold, int sellPriceByDiamond, int minLevel, bool sellable, bool tradable, bool destroyable,
         string description, string sprite, bool useDestroy, int useHealth, int useMagic, int useExperience,
         int equipHealthBonus, int equipManaBonus, int equipDamageBonus, int equipDefenseBonus, int equipSpeedcBonus,
-        GameObject modelPrefab, Drawing[] drawing,int drawingItemId,float makeTime)
+        GameObject modelPrefab, Drawing[] drawing,int drawingItemId,int makeTime)
     {
         this.id = id;
         this.itemName = itemName;
